refactor: projectService interface 분리
